feat(app-vite): split types generation, handle it on diff now it will re-generate the types if build.typescript or build.alias gets updated
